requirements.txt
pandas==2.1.4
pydantic==2.5.3
pytest==7.4.3
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
